🔄 Add CombineOnSuccessWith methods for IResults
We need a way to combine success results, collecting all their values as
we go. If an error happens anywhere in the stack, we would like to not
execute any remaining ❌🤐🔙 and instead instantly return the failure.

The added methods recursively work through a list, lazily evaluating the
❌🤐🔙 based on a function provided, and aggregating the result
(unless an error occurs).


🔄 Original Commit: https://github.com/foxguardsolutions/Incipience/commit/415bd0d98252ac2c7071d5852845b7a70314eac2