🔄 Persist ❌🤐🔙 results to a file on ❌🤐🔙
Previously ❌🤐🔙 results were simply read into memory and formatted to
correctly display on an HTML page. We now want to save them to a file in
order to enable ❌🤐🔙 on the files later.

This commit introduces a way to save ❌🤐🔙 results to a file as a single
string and manipulates that string accordingly for viewing.


🔄 Original Commit: https://github.com/foxguardsolutions/Incipience/commit/26f06512cfc202b06cbf04200ae3edbc8eedd738