🔄 DRYs Common Project Properties
MSBuild recently added the ability to [share common project properties](https://www.thomaslevesque.com/2017/09/18/common-msbuild-properties-and-items-with-directory-build-props/) in a root file named & discovered by convention. This leverages such in order to reduce the burden of adding new projects and also to make it easier to amend common properties in the future.

Includes the common properties file (and a few other root-of-the-solution files into the solution for easy access in Visual Studio.

Updates the editor config to make editors behave a bit better when editing project files.


🔄 Original Commit: https://github.com/foxguardsolutions/Incipience/commit/da274d7c277104992854f845fb964d933a94d3f7